Fix mypy errors

+++ packages/syft/src/syft/core/tensor/autograd/backward_ops/abs.py
@@ -19,7 +19,7 @@ class AbsOp(Op):
 
         if self.x.requires_grad:
             _grad = self.x > 0  # returns 0s and 1s
-            _grad = (_grad * 2) - 1  # returns -1s and 1s
+            _grad = int((_grad * 2) - 1)  # returns -1s and 1s
 
             grad = _grad * grad
